+++ swarm/CHANGELOG.md
@@ -11,6 +11,8 @@
 - Don't close entire connection upon `DialUpgradeError`s within `OneShotHandler`.
   Instead, the error is reported as `Err(e)` via `ConnectionHandler::ToBehaviour`.
   See [PR 4715](https://github.com/libp2p/rust-libp2p/pull/4715).
+- Add `SwarmEvent::{NewExternalAddrCandidate,ExternalAddrConfirmed,ExternalAddrExpired}` variants.
+  See [PR XXXX](https://github.com/libp2p/rust-libp2p/pull/XXXX).
 
 ## 0.43.6